New item listed in local repository

+++ c/c99/1002_area_do_circulo.c
@@ -2,12 +2,12 @@
 #include <math.h>
 
 int main() {
-    float raio, area, n = 3.14159;
+    double raio, area, n = 3.14159;
 
-    scanf("%f", &raio);
+    scanf("%lf", &raio);
     area = n * pow(raio, 2);
 
-    printf("A=%.4f\n", area);
+    printf("A=%.4lf\n", area);
 
     return 0;
 }
